# Conflicts:
#	src/main/java/com/example/demo1/Anamnese.java
#	src/main/java/com/example/demo1/App.java
#	src/main/java/com/example/demo1/QCM.java
#	src/main/resources/com/example/demo1/QCM.txt
#	src/main/resources/com/example/demo1/bilan.fxml

+++ src/main/java/com/example/demo1/Consultation.java
@@ -14,7 +14,7 @@ public class Consultation extends RendezVous {
         this.age=age;
         this.type=TypeRdv.CONSULTATION;
     }
-    public void programmer(LocalDate dat, LocalTime heure, float duree) {
+    public void programmer(LocalDate dat, LocalTime heure, int duree) {
         this.dat=dat;
         this.time=heure;
         this.type=TypeRdv.CONSULTATION;
